# Optimization: Add an optimized path for notifying property changes for inherited va…