Setting DataContext or other inherited properties will allocate many instances of AvaloniaPropertyChangedEventArgs and PropertyChangedEventArgs, an instance on every node in the tree that it raises on.

This PR reuses the same instance.

This situation wont really be a problem in most applications, but can become an issue with heavy updates of cells when using DataGrid or TreeDataGrid.


Benchmark:

|            Method |     Mean |    Error |   StdDev |       Gen0 | Allocated |
|------------------ |---------:|---------:|---------:|-----------:|----------:|
| ChangeDataContext | 412.2 ms | 17.16 ms | 50.05 ms | 18000.0000 | 150.01 MB |

|            Method |     Mean |   Error |  StdDev | Allocated |
|------------------ |---------:|--------:|--------:|----------:|
| ChangeDataContext | 241.4 ms | 4.02 ms | 3.76 ms |  54.52 KB |